Stability diagram of the Bowtie Cavity
Numerical calculation of the stability diagram of the bowtie cavity proposed in the section "26.04.2013 Dimensioning the ring cavity" of the LogBook. I assume a fixed laser beam geometry, where the round trip length is 320.8mm, the distance between the spherical mirrors is 85mm, and the distance between the plane mirrors is 75mm. I also assume the available concave mirrors on the Thorlabs website (https://www.thorlabs.com/newgrouppage9.cfm?objectgroup_id=6636).